Show all questions
The user can see all the questions after a function is called.

Added test data for the quiz in an array of objects.

Added a showQuiz() function that dynamically creates each question.

Next thing to do is to dynamically create the answer options.

File krav.md: Updated the requirements to divide the development process in smaller steps.