Show answer options dynamically
The user can see all the options for each question. Only one answer per question can be selected.

The app dynamically generates the options by iterating through all the options for each question in the quiz. I used a for...in statement to achieve this.

Each option has a radio button. I decided to put the input tag inside the label tag. Doing it this way makes the text in the option clickable, not just the radio button.