Show checkboxes for multiple answer questions
The user can select more than one correct answer for some questions.

We have two types of questions: ones that only have one correct answer and ones that have more than one correct answer. To check what type it is, I used a conditional that creates a checkbox for multiple correct answers and a radio button for single correct answer.

All the correct answers for multiple types are saved in an array, so I decided to use that in the conditional. In other words, it checks if the answers are in an array.