Check answers with a function
The app checks if the user's answers are correct when a function is run.

This commit will add the new function checkAnswers(). It compares the results from the quiz that has all the correct answers, against the user selected answers.

Since the app takes in multiple correct answers and single correct answers, it had to have a way to check both cases.

To achieve this I added a dedicated check for multiple correct answers since they are saved in an array. The problem with comparing whether two arrays are equal is tricky because you get false even though their values are the same.

To be able to compare two separate arrays, I decided to use the JSON.stringify method. This converts the array to a string. I thought this was the best option for this case because it's short and clean, as well as it handles cases where not all correct options are selected or if more than the correct option are selected.