Delete score message when redoing quiz
When the user has finished answering the quiz for the first time, a message with the score information shows up. But when the Redo Quiz button is clicked, the message didn't clear.

To avoid duplicating the same code for removing the message that was already in the showResult() function, I decided to move it to its own function deleteScoreMessage().
This new function is used in showResult() and restartQuiz().