Allow to check answers only if quiz is completed
There is a message at the bottom of the quiz that asks the user to answer all questions. When the user has answered all questions, the Check Answers button will show up and the message will disappear.

This is achieved with the new function isQuizComplete(). The app checks if the quiz is complete every time there’s an answer is selected.