Count total answered questions
The app was showing the button to check the score when a user selected at least 10 options. This button should only be visible when all questions are answered.

The app assumed that each question had one single selected answer, but that’s wrong. We have questions were multiple options can be selected. Counting the number of selected options was not an accurate way to check how many questions were answered. As a result, the button showed up even though there were unanswered questions left.

A counter to check the total answered questions solves this problem. The app goes through each question to find if there is at least one selected answer, if that’s the case it adds it to the total.

This commit will update the isQuizComplete() function.