Remove duplicate saveAnswers() function call
The app was saving the answers twice: when all the questions were answered and when the user clicked the button to check the score. We only need to save the answers once.

It is unnecessary to save the answers when all the questions are answered. The user might decide to change an answer.

I decided to leave it in the checkAnswers() function so the selected answers are saved only when the user is ready to check the score.