Select quiz inputs after they are dynamically created
The app was not showing the button to check the result after all questions were answered.

This bug happened when the showQuiz() function was moved so it’s called with the Start Quiz button.

The isQuizComplete() function was not running because the app was looking for the input elements before they were dynamically created.

To solve this I added a conditional that checks if quiz inputs exist.